add Kill dtos validation

+++ src/Models/Dtos/Kill/KillDto.cs
@@ -10,7 +10,7 @@ namespace AirsoftBattlefieldManagementSystemAPI.Models.Dtos.Kill
         public int RoomId { get; set; }
         public decimal Longitude { get; set; }
         public decimal Latitude { get; set; }
-        public short Accuracy { get; set; }
+        public decimal Accuracy { get; set; }
         public short Bearing { get; set; }
         public DateTime Time { get; set; }
     }
